Separated vtm-playground/org.oscim.test.gdx.poi3d to vtm-gdx-models
