+++ vtm-gdx-models/src/org/oscim/model/poi3d/GdxModelRenderer.java
@@ -1,4 +1,4 @@
-package org.oscim.test.gdx.poi3d;
+package org.oscim.model.poi3d;
 
 import com.badlogic.gdx.graphics.g3d.Environment;
 import com.badlogic.gdx.graphics.g3d.Model;
@@ -8,7 +8,6 @@ import com.badlogic.gdx.graphics.g3d.Renderable;
 import com.badlogic.gdx.graphics.g3d.Shader;
 import com.badlogic.gdx.graphics.g3d.attributes.ColorAttribute;
 import com.badlogic.gdx.graphics.g3d.environment.DirectionalLight;
-import com.badlogic.gdx.graphics.g3d.shaders.DefaultShader;
 import com.badlogic.gdx.graphics.g3d.utils.DefaultShaderProvider;
 import com.badlogic.gdx.graphics.g3d.utils.DefaultTextureBinder;
 import com.badlogic.gdx.graphics.g3d.utils.RenderContext;
@@ -22,14 +21,13 @@ import org.oscim.map.Viewport;
 import org.oscim.renderer.GLState;
 import org.oscim.renderer.GLViewport;
 import org.oscim.renderer.LayerRenderer;
-import org.oscim.utils.geom.GeometryUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static org.oscim.backend.GLAdapter.gl;
 
-public class GdxRenderer3D extends LayerRenderer {
-    static final Logger log = LoggerFactory.getLogger(GdxRenderer3D.class);
+public class GdxModelRenderer extends LayerRenderer {
+    static final Logger log = LoggerFactory.getLogger(GdxModelRenderer.class);
 
     ModelBatch modelBatch;
     public MapCamera cam;
@@ -44,8 +42,9 @@ public class GdxRenderer3D extends LayerRenderer {
     public Shader shader;
     public RenderContext renderContext;
     public Model model;
+    private ModelBatch mBatch = new ModelBatch();
 
-    public GdxRenderer3D(Map map) {
+    public GdxModelRenderer(Map map) {
         mMap = map;
     }
 
@@ -99,8 +98,8 @@ public class GdxRenderer3D extends LayerRenderer {
 
         gl.depthMask(true);
 
-        // if (position.zoomLevel < 17)
-        // GL.clear(GL20.DEPTH_BUFFER_BIT);
+        if (v.pos.zoomLevel < 16)
+            gl.clear(GL.DEPTH_BUFFER_BIT);
 
         gl.bindBuffer(GL.ELEMENT_ARRAY_BUFFER, 0);
 
@@ -113,10 +112,6 @@ public class GdxRenderer3D extends LayerRenderer {
         GLState.blend(false);
 
         cam.update(v);
-        long time = System.currentTimeMillis();
-
-        int cnt = 0;
-        int rnd = 0;
 
         Viewport p = mMap.viewport();
         p.getMapExtents(mBox, 10);
@@ -134,56 +129,47 @@ public class GdxRenderer3D extends LayerRenderer {
             mBox[i + 1] -= dy;
         }
 
-        int w = mMap.getWidth() / 2;
-        int h = mMap.getHeight() / 2;
-
-        float sqRadius = (w * w + h * h) / scale;
+        //int w = mMap.getWidth() / 2;
+        //int h = mMap.getHeight() / 2;
+        //float sqRadius = (w * w + h * h) / scale;
 
         synchronized (this) {
             if (instances.size == 0)
                 return;
 
-            cnt = instances.size;
+            //renderContext.begin();
 
-            renderContext.begin();
-
-            if (shader == null) {
-                r = instances.get(0).getRenderable(r);
-                DefaultShader.Config c = new DefaultShader.Config();
-                c.numBones = 0;
-                c.numDirectionalLights = 1;
-                r.environment = lights;
-                // shader = new DefaultShader(r, true, false, false, false, 1,
-                // 0, 0, 0);
-                shader = new DefaultShader(r, c);
-                shader.init();
-            }
-
-            shader.begin(cam, renderContext);
+            //            if (shader == null) {
+            //                r = instances.get(0).getRenderable(r);
+            //                DefaultShader.Config c = new DefaultShader.Config();
+            //                c.numBones = 0;
+            //                c.numDirectionalLights = 1;
+            //                r.environment = lights;
+            //
+            //                shader = new DefaultShader(r, c);
+            //                shader.init();
+            //            }
+            mBatch.begin(cam);
+            //shader.begin(cam, renderContext);
 
             for (ModelInstance instance : instances) {
                 instance.transform.getTranslation(tempVector);
+                //instance.getRenderables(renderables, pool);
+                //    if (tempVector.x * tempVector.x + tempVector.y * tempVector.y > sqRadius)
+                //     continue;
+                //    tempVector.scl(0.8f, 0.8f, 1);
+                //    if (!GeometryUtils.pointInPoly(tempVector.x, tempVector.y, mBox, 8, 0))
+                //    continue;
 
-                if (tempVector.x * tempVector.x + tempVector.y * tempVector.y > sqRadius)
-                    continue;
+                mBatch.render(instance);
 
-                tempVector.scl(0.8f, 0.8f, 1);
-
-                if (!GeometryUtils.pointInPoly(tempVector.x, tempVector.y, mBox, 8, 0))
-                    continue;
-
-                instance.getRenderable(r);
-                // r.lights = lights;
-                // r.environment = lights;
-                shader.render(r);
-
-                rnd++;
+                //shader.render(r);
             }
+            mBatch.end();
 
-            shader.end();
-            renderContext.end();
+            //shader.end();
+            //renderContext.end();
         }
-        log.debug(">>> " + (System.currentTimeMillis() - time) + " " + cnt + "/" + rnd);
 
         gl.depthMask(false);
         gl.bindBuffer(GL.ELEMENT_ARRAY_BUFFER, 0);
