+++ vtm-gdx-models/src/org/oscim/model/poi3d/Poi3DLayer.java
@@ -14,7 +14,7 @@
  * this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.oscim.test.gdx.poi3d;
+package org.oscim.model.poi3d;
 
 import com.badlogic.gdx.assets.AssetManager;
 import com.badlogic.gdx.graphics.g3d.Model;
